rename signal, fix import and typos 1
